requirements.txt

# Phase 4.3 Enhanced Video Creation Dependencies  
moviepy>=1.0.3
scipy>=1.11.0
librosa>=0.10.0
pydub>=0.25.1

# Advanced audio processing
ffmpeg-python>=0.2.0
soundfile>=0.12.1
mutagen>=1.47.0

# Existing Phase 4.2 dependencies
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Core Azure and AI dependencies
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-cognitiveservices-speech>=1.34.0
azure-ai-translation-text>=1.0.0
openai>=1.6.1
fastapi>=0.104.1
requests>=2.31.0
